Refactor search code to new crate
